feat: Added shadow bias
A possible way of dealing with shadow acne, the bias offsets the
intersection point to slightly above the object along the normal.
However, the floating point calculations up to cause the margin of
error to be quite large, meaning a larger shadow bias is necessary.
This causes issues to shadowcasting. More testing necessary.